Initial Source RIP
Attempted to use dotPeek to RIP the files and it would ahve worked but only for scripts. Instead i decided to use a dedicated Unity tool i came across called Utiny Ripper which retrieved majority of the files (including scenes and textures) automatically and set them up in the normal Unity folder structure.